ref: Move utils from raven.js to utils.js file (#1084)
